engine: extract dispatch::phases (#159)
