+++ crates/game-core/src/engine/dispatch/phases.rs
@@ -1,0 +1,2345 @@
+//! Phase-driver functions: start/end scenario, per-phase entrypoints,
+//! and the round-cycle stepping logic.
+
+use crate::engine::outcome::EngineOutcome;
+use crate::event::Event;
+use crate::state::{EnemyId, InvestigatorId, Phase, WindowKind};
+
+/// Action points granted to an investigator at the start of their
+/// turn during the Investigation phase. Per the Arkham Horror LCG
+/// rulebook.
+pub(super) const ACTIONS_PER_TURN: u8 = 3;
+
+pub(super) fn start_scenario(
+    state: &mut crate::state::GameState,
+    events: &mut Vec<Event>,
+) -> EngineOutcome {
+    // The GameState constructor places the world in its initial shape;
+    // this action is the explicit "session has begun" marker that lands
+    // in the action log. Replaying it on an already-started state is a
+    // bug, not a no-op — reject so callers notice rather than silently
+    // double-emitting `ScenarioStarted`.
+    if state.round != 0 {
+        return EngineOutcome::Rejected {
+            reason: "StartScenario applied to a state that is already in progress".into(),
+        };
+    }
+    // Round 1: scenario starts directly in Investigation phase —
+    // Mythos is skipped entirely per Rules Reference p.24 "During
+    // the first round of the game, skip the mythos phase." No
+    // PhaseStarted(Mythos) / PhaseEnded(Mythos) fire — the phase
+    // doesn't happen.
+    state.round = 1;
+    state.phase = Phase::Investigation;
+    events.push(Event::ScenarioStarted);
+
+    // For each investigator (sorted by id for determinism), shuffle
+    // their deck and deal an initial hand of up to 5.
+    let inv_ids: Vec<InvestigatorId> = state.investigators.keys().copied().collect();
+    for inv_id in inv_ids {
+        super::cards::shuffle_player_deck(state, events, inv_id);
+        super::cards::draw_cards(state, events, inv_id, super::cards::INITIAL_HAND_SIZE);
+    }
+
+    // Seed the mulligan cursor to the first Active investigator in
+    // player order. Each investigator submits a single
+    // `PlayerAction::Mulligan` in turn; the cursor advances after each
+    // and reaches `None` once all have gone (see `apply_player_action`),
+    // at which point setup ends. Other player actions are rejected while
+    // the cursor is `Some`. An empty/all-eliminated `turn_order` seeds
+    // `None` — the same degenerate no-op as the Mythos draw cursor.
+    state.mulligan_pending = super::cursor::first_active_investigator(state);
+
+    // Round-1 action seed: round 1 skips Mythos, so there's no Upkeep 4.2
+    // to grant the first round's actions. Every Active investigator → ACTIONS_PER_TURN.
+    reset_actions(state, events);
+
+    // NOTE: the Investigation phase is NOT begun here. Setup has no
+    // action windows (Rules Reference p.27); the phase begins after the
+    // mulligan cursor reaches `None` — see the kickoff in apply_player_action.
+    EngineOutcome::Done
+}
+
+pub(super) fn end_turn(
+    state: &mut crate::state::GameState,
+    events: &mut Vec<Event>,
+) -> EngineOutcome {
+    if state.phase != Phase::Investigation {
+        return EngineOutcome::Rejected {
+            reason: "EndTurn is only valid during the Investigation phase".into(),
+        };
+    }
+    let Some(active_id) = state.active_investigator else {
+        return EngineOutcome::Rejected {
+            reason: "EndTurn requires an active investigator".into(),
+        };
+    };
+    // The Some(active_investigator) invariant is paired with that ID
+    // existing in the investigators map; a missing entry would be state
+    // corruption, not a normal rejection. Surface it loudly rather than
+    // hiding behind Rejected.
+    let active = state.investigators.get_mut(&active_id).unwrap_or_else(|| {
+        unreachable!(
+            "active_investigator {active_id:?} is not in the investigators map; \
+                 this is a state-corruption invariant violation"
+        )
+    });
+
+    // Drain remaining actions and announce the turn ended.
+    if active.actions_remaining != 0 {
+        active.actions_remaining = 0;
+        events.push(Event::ActionsRemainingChanged {
+            investigator: active_id,
+            new_count: 0,
+        });
+    }
+    events.push(Event::TurnEnded {
+        investigator: active_id,
+    });
+
+    // 2.2.2 decision: "return to 2.2" for the next investigator, or
+    // proceed to 2.3. next_active_investigator_after skips eliminated
+    // investigators (Rules Reference p.10) — the same shared helper the
+    // Enemy phase uses.
+    if let Some(next_id) = super::cursor::next_active_investigator_after(state, active_id) {
+        begin_investigator_turn(state, events, next_id);
+        EngineOutcome::Done
+    } else {
+        state.active_investigator = None;
+        // 2.3 → Enemy. The cascade may suspend on a hunter-movement tie
+        // (Enemy 3.2); propagate its outcome rather than swallowing it.
+        investigation_phase_end(state, events)
+    }
+}
+
+/// Entered by [`step_phase`] on any-to-Investigation transition, and by
+/// the mulligan-completion site in [`apply_player_action`] for round 1.
+/// Owns the `PhaseStarted(Investigation)` emit (Rules Reference p.24
+/// step 2.1) and opens the post-2.1 player window. Rotation to the
+/// first active investigator (step 2.2) runs in the
+/// [`WindowKind::InvestigationBegins`] continuation via
+/// [`begin_investigator_turn`], lead-first by default; explicit
+/// player-pick within this window is deferred to #146.
+///
+/// The window auto-skips inline when nothing is Fast-eligible
+/// ([`any_fast_play_eligible`] returns `false` — e.g. no Fast card in any
+/// hand, which is always the case in unit tests with no card registry
+/// installed), so single-investigator entry still lands the lead active
+/// within the same `apply()` call.
+pub(super) fn investigation_phase(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    // 2.1 Investigation phase begins.
+    events.push(Event::PhaseStarted {
+        phase: Phase::Investigation,
+    });
+    // PLAYER WINDOW (post-2.1). Rotation to the first investigator
+    // (step 2.2) runs in this window's continuation
+    // (`run_window_continuation` → `InvestigationBegins`), so the printed
+    // order 2.1 → window → 2.2 holds. Auto-skips inline when nothing is
+    // Fast-eligible, so single-investigator entry still lands the lead
+    // active within the same apply() call.
+    super::reaction_windows::open_fast_window(state, events, WindowKind::InvestigationBegins);
+}
+
+/// 2.2 Next investigator's turn begins. Rotates the active cursor to
+/// `who` (the chosen/default investigator) and opens the post-2.2
+/// player window. Called from the `InvestigationBegins` continuation
+/// (first turn of the phase) and from `end_turn` (each subsequent turn,
+/// the rules' "return to 2.2"). Step
+/// 2.2.1 (the active investigator's actions) follows as player-driven
+/// inputs while `InvestigatorTurnBegins` is the "previous player window."
+///
+/// `who` must be an `Active` investigator in `turn_order`; callers
+/// resolve it via `first_active_investigator` / `next_active_investigator_after`.
+pub(super) fn begin_investigator_turn(
+    state: &mut crate::state::GameState,
+    events: &mut Vec<Event>,
+    who: InvestigatorId,
+) {
+    rotate_to_active(state, events, who);
+    super::reaction_windows::open_fast_window(state, events, WindowKind::InvestigatorTurnBegins);
+}
+
+/// 2.3 Investigation phase ends. Owns the `PhaseEnded(Investigation)`
+/// emit — lifted out of `step_phase`, mirroring `mythos_phase_end` /
+/// `enemy_phase_end` / `upkeep_phase_end` — then transitions to the
+/// Enemy phase. Called only from `end_turn`'s terminal branch (the last
+/// investigator has taken a turn this round).
+fn investigation_phase_end(
+    state: &mut crate::state::GameState,
+    events: &mut Vec<Event>,
+) -> EngineOutcome {
+    events.push(Event::PhaseEnded {
+        phase: Phase::Investigation,
+    });
+    step_phase(state, events) // Investigation → Enemy; calls enemy_phase
+}
+
+/// Entered by [`step_phase`] on the Upkeep→Mythos transition. Lays
+/// out the Rules Reference p.24 sub-steps as discrete named call
+/// sites so the rule structure is grep-able and #73 / future-peril-PR
+/// fills in TODO bodies without changing the driver shape.
+fn mythos_phase(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    // 1.1 Round begins. Mythos phase begins.
+    //     Rules Reference p.24: "As this is the first framework event
+    //     of the round, it [1.1] also formalizes the beginning of a new
+    //     game round." The round-counter increment lives HERE (not in
+    //     step_phase) so the rule's round-begin point has explicit
+    //     driver ownership, mirroring PhaseStarted(Mythos). Round 1 is
+    //     bypassed: start_scenario sets round = 1 directly (Mythos
+    //     skipped). This is also the future home for a RoundStarted
+    //     event when a consumer lands.
+    state.round = state.round.saturating_add(1);
+    events.push(Event::PhaseStarted {
+        phase: Phase::Mythos,
+    });
+
+    // 1.2 Place 1 doom on the current agenda.
+    super::act_agenda::place_doom_on_agenda(state, events);
+
+    // 1.3 Check doom threshold.
+    super::act_agenda::check_doom_threshold(state, events);
+
+    // 1.4 Each investigator draws 1 encounter card.
+    //     Seed the cursor; the actual draws are player-driven via
+    //     PlayerAction::DrawEncounterCard (lands in T12). The
+    //     dispatch handler advances the cursor after each chain.
+    //     Per Rules Reference p.10 (Elimination), eliminated
+    //     investigators (Killed, Insane, Resigned) do not draw
+    //     encounter cards — skip to the first Active investigator.
+    state.mythos_draw_pending = super::cursor::first_active_investigator(state);
+    if state.mythos_draw_pending.is_none() {
+        // No Active investigators to draw (turn_order is empty or all
+        // investigators are eliminated). Open the post-1.4 window
+        // immediately; open_fast_window's auto-skip path triggers
+        // because nothing is eligible, runs the MythosAfterDraws
+        // continuation (mythos_phase_end), which transitions to
+        // Investigation. All in this same apply.
+        super::reaction_windows::open_fast_window(state, events, WindowKind::MythosAfterDraws);
+    }
+}
+
+/// Transition to the next phase. Dispatches into phase driver
+/// functions when they exist (each driver owns its own
+/// `PhaseStarted` emit). For phases without a driver, emits
+/// `PhaseStarted` directly.
+///
+/// **`PhaseEnded` invariant:** `step_phase` emits **no** `PhaseEnded`
+/// for any phase. Each phase's `*_end` helper owns its own boundary
+/// emit: `mythos_phase_end` (step 1.5), `investigation_phase_end`
+/// (step 2.3), `enemy_phase_end` (step 3.4), `upkeep_phase_end`
+/// (step 4.6). `start_scenario`'s first-round-skip path bypasses the
+/// entire Mythos phase — no `PhaseStarted(Mythos)` /
+/// `PhaseEnded(Mythos)` events fire on round 1 — per Rules Reference
+/// p.24 ("skip the mythos phase").
+///
+/// **Round-bump:** the round-counter increment now lives in
+/// `mythos_phase` step 1.1 — the rules' "round begins" point —
+/// rather than here. `step_phase` no longer touches `state.round`.
+///
+/// Returns the transition's [`EngineOutcome`]. Only the
+/// Investigation→Enemy arm can return [`EngineOutcome::AwaitingInput`]
+/// (a hunter-movement tie in [`enemy_phase`]); every other arm runs its
+/// driver to completion and returns [`EngineOutcome::Done`]. The
+/// Investigation→Enemy suspension is owned by
+/// [`investigation_phase_end`], which propagates it up through
+/// [`end_turn`].
+fn step_phase(state: &mut crate::state::GameState, events: &mut Vec<Event>) -> EngineOutcome {
+    let from = state.phase;
+    let to = from.next();
+
+    state.phase = to;
+    // The round-counter bump moves into mythos_phase (step 1.1).
+    // step_phase no longer touches state.round.
+
+    // Dispatch to phase driver if one exists; otherwise emit
+    // PhaseStarted directly (for phases without a driver yet).
+    match to {
+        Phase::Mythos if from != Phase::Mythos => {
+            mythos_phase(state, events);
+            EngineOutcome::Done
+        }
+        Phase::Investigation if from != Phase::Investigation => {
+            investigation_phase(state, events);
+            EngineOutcome::Done
+        }
+        Phase::Enemy if from != Phase::Enemy => enemy_phase(state, events),
+        Phase::Upkeep if from != Phase::Upkeep => {
+            upkeep_phase(state, events);
+            EngineOutcome::Done
+        }
+        _ => unreachable!(
+            "step_phase: from == to (from={from:?}, to={to:?}); Phase::next \
+             never returns the same phase, so this branch is structurally \
+             unreachable. If it ever fires, something has corrupted \
+             state.phase between the read and the dispatch."
+        ),
+    }
+}
+
+/// Set `active_investigator` to `id`. Does NOT refresh actions —
+/// actions are reset at Upkeep step 4.2 (`reset_actions`) for the whole
+/// next round, and seeded for round 1 by `start_scenario`. By the time
+/// an investigator becomes active, `actions_remaining` already holds
+/// this round's allotment.
+///
+/// `id` must refer to an investigator in `state.investigators` (a
+/// whole-program invariant for ids drawn from `turn_order`).
+fn rotate_to_active(
+    state: &mut crate::state::GameState,
+    _events: &mut Vec<Event>,
+    id: InvestigatorId,
+) {
+    debug_assert!(
+        state.investigators.contains_key(&id),
+        "rotate_to_active: investigator {id:?} not in investigators (state corruption)"
+    );
+    state.active_investigator = Some(id);
+}
+
+/// 3.3 Seed the per-investigator attack cursor and open the first
+/// attack window — or the final window directly if there is no Active
+/// investigator. Called once hunter movement (step 3.2) completes:
+/// from [`enemy_phase`] on the no-tie path, and from
+/// [`resume_hunter_choice`] once all hunters resolve.
+///
+/// Seeds the cursor to the first Active investigator in `turn_order`.
+/// Eliminated investigators (Killed / Insane / Resigned) are skipped per
+/// Rules Reference p.10 (Elimination); [`cursor::first_active_investigator`] is
+/// the shared helper used by Mythos 1.4 (#69) for the same semantics.
+/// The loop body runs in [`run_window_continuation`]'s arms.
+pub(super) fn enemy_attack_kickoff(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    state.enemy_attack_pending = super::cursor::first_active_investigator(state);
+
+    if state.enemy_attack_pending.is_some() {
+        super::reaction_windows::open_fast_window(
+            state,
+            events,
+            WindowKind::BeforeInvestigatorAttacked,
+        );
+    } else {
+        // No Active investigators (turn_order empty or all eliminated).
+        // Skip straight to the final window — mirror of mythos_phase's
+        // no-drawer path.
+        super::reaction_windows::open_fast_window(
+            state,
+            events,
+            WindowKind::AfterAllInvestigatorsAttacked,
+        );
+    }
+}
+
+/// Entered by [`step_phase`] on the Investigation→Enemy transition.
+/// Owns the `PhaseStarted(Enemy)` emit (Rules Reference p.25 step 3.1),
+/// runs hunter movement (step 3.2) via [`drive_hunter_moves`], then
+/// kicks off the per-investigator attack loop (step 3.3) via
+/// [`enemy_attack_kickoff`].
+///
+/// If hunter movement suspends on a lead-investigator tie, this returns
+/// the [`EngineOutcome::AwaitingInput`] unchanged — the attack-loop
+/// kickoff is deferred to [`resume_hunter_choice`], which runs it once
+/// the last hunter resolves. Otherwise the kickoff runs inline here and
+/// this returns [`EngineOutcome::Done`].
+fn enemy_phase(state: &mut crate::state::GameState, events: &mut Vec<Event>) -> EngineOutcome {
+    // 3.1 Enemy phase begins.
+    events.push(Event::PhaseStarted {
+        phase: Phase::Enemy,
+    });
+
+    // 3.2 Hunter enemies move. Park on a lead-investigator tie; the
+    //     attack-loop kickoff then happens on resume.
+    match super::hunters::drive_hunter_moves(state, events) {
+        outcome @ EngineOutcome::AwaitingInput { .. } => return outcome,
+        // drive_hunter_moves only ever returns Done or AwaitingInput, never Rejected.
+        EngineOutcome::Rejected { reason } => {
+            unreachable!("enemy_phase: hunter movement rejected unexpectedly: {reason}")
+        }
+        EngineOutcome::Done => {}
+    }
+
+    // 3.3 Kick off the per-investigator attack loop.
+    enemy_attack_kickoff(state, events);
+    EngineOutcome::Done
+}
+
+/// Called from [`run_window_continuation`]'s
+/// [`WindowKind::AfterAllInvestigatorsAttacked`] arm. Emits step
+/// 3.4's `PhaseEnded(Enemy)` marker, then transitions to Upkeep.
+/// Exact analog of [`mythos_phase_end`] / [`upkeep_phase_end`].
+pub(super) fn enemy_phase_end(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    // 3.4 Enemy phase ends.
+    events.push(Event::PhaseEnded {
+        phase: Phase::Enemy,
+    });
+    // Enemy → Upkeep; calls upkeep_phase. Only the Investigation→Enemy
+    // transition can suspend (hunter movement), so this never does.
+    let outcome = step_phase(state, events);
+    debug_assert_eq!(
+        outcome,
+        EngineOutcome::Done,
+        "unexpected suspension in Enemy→Upkeep transition"
+    );
+}
+
+/// Called after the post-1.4 window closes. Emits 1.5's
+/// `PhaseEnded(Mythos)` marker, then transitions to Investigation.
+/// Rotation is owned by `investigation_phase` (step 2.2), not by
+/// `mythos_phase_end`. Invoked from `close_reaction_window_at`'s
+/// kind-aware tail when a `MythosAfterDraws` window pops, and from
+/// `open_fast_window`'s auto-skip path inline.
+pub(super) fn mythos_phase_end(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    // 1.5 Mythos phase ends.
+    //     The PhaseEnded(Mythos) emit lives HERE rather than in
+    //     step_phase so step 1.5 has explicit ownership in the
+    //     driver — mirror of step 1.1's PhaseStarted ownership in
+    //     mythos_phase. Rules Reference p.24: "This step formalizes
+    //     the end of the mythos phase."
+    events.push(Event::PhaseEnded {
+        phase: Phase::Mythos,
+    });
+    // Mythos → Investigation; calls investigation_phase. Only the
+    // Investigation→Enemy transition can suspend (hunter movement), so
+    // this cascade always completes.
+    let outcome = step_phase(state, events);
+    debug_assert_eq!(
+        outcome,
+        EngineOutcome::Done,
+        "unexpected suspension in Mythos→Investigation transition"
+    );
+}
+
+/// Entered by [`step_phase`] on the Enemy→Upkeep transition. Owns the
+/// `PhaseStarted(Upkeep)` emit (step 4.1) and opens the post-4.1 player
+/// window. Steps 4.2 onward run as the window's continuation
+/// ([`upkeep_resume`]). Mirror of [`mythos_phase`], inverted: Mythos's
+/// window sits at the END, so its driver runs content then opens;
+/// Upkeep's sits at the START, so the driver opens immediately and the
+/// content is the continuation.
+fn upkeep_phase(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    // 4.1 Upkeep phase begins.
+    events.push(Event::PhaseStarted {
+        phase: Phase::Upkeep,
+    });
+    // PLAYER WINDOW (post-4.1). Auto-skips inline (running upkeep_resume
+    // via run_window_continuation) when nothing is Fast-eligible.
+    super::reaction_windows::open_fast_window(state, events, WindowKind::UpkeepBegins);
+}
+
+/// The post-4.1 window continuation. Steps 4.2–4.4 run inline as named
+/// call sites; step 4.5 is the [`check_hand_size`] stub (TODO #111).
+/// Then hands to [`upkeep_phase_end`] for 4.6 + transition.
+pub(super) fn upkeep_resume(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    reset_actions(state, events); // 4.2
+    ready_exhausted_cards(state, events); // 4.3
+    upkeep_draw_and_resource(state, events); // 4.4
+    check_hand_size(state, events); // 4.5 (TODO #111)
+    upkeep_phase_end(state, events); // 4.6 + transition
+}
+
+/// Owns step 4.6's `PhaseEnded(Upkeep)` emit, then transitions to
+/// Mythos. Exact analog of [`mythos_phase_end`]. `step_phase` emits no
+/// `PhaseEnded` itself — every phase's `*_end` helper owns its own.
+fn upkeep_phase_end(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    // 4.6 Upkeep phase ends. Round ends.
+    events.push(Event::PhaseEnded {
+        phase: Phase::Upkeep,
+    });
+    // Upkeep → Mythos; calls mythos_phase. Only the Investigation→Enemy
+    // transition can suspend (hunter movement), so this never does.
+    let outcome = step_phase(state, events);
+    debug_assert_eq!(
+        outcome,
+        EngineOutcome::Done,
+        "unexpected suspension in Upkeep→Mythos transition"
+    );
+}
+
+/// 4.3 Ready exhausted cards. Rules Reference p.25: "Simultaneously
+/// ready each exhausted card." "Each exhausted card" is every exhausted
+/// card in play regardless of controller — investigator in-play cards
+/// AND enemies. Simultaneous, so iteration order is immaterial; we
+/// iterate deterministically (investigator id then in-play order; then
+/// enemy id) for reproducible event streams. Already-ready cards emit
+/// nothing.
+///
+/// After readying, each enemy that became ready while unengaged and
+/// co-located with an investigator engages it via [`reengage_at_location`]
+/// (Rules Reference p.10: "if an exhausted enemy at the same location as an
+/// investigator becomes ready, it engages as soon as it is readied").
+fn ready_exhausted_cards(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    let inv_ids: Vec<InvestigatorId> = state.investigators.keys().copied().collect();
+    for id in inv_ids {
+        let inv = state.investigators.get_mut(&id).expect("id from keys");
+        for card in &mut inv.cards_in_play {
+            if card.exhausted {
+                card.exhausted = false;
+                events.push(Event::CardReadied {
+                    investigator: id,
+                    instance_id: card.instance_id,
+                    code: card.code.clone(),
+                });
+            }
+        }
+    }
+    let enemy_ids: Vec<EnemyId> = state.enemies.keys().copied().collect();
+    let mut newly_readied: Vec<EnemyId> = Vec::new();
+    for eid in enemy_ids {
+        let enemy = state.enemies.get_mut(&eid).expect("id from keys");
+        if enemy.exhausted {
+            enemy.exhausted = false;
+            events.push(Event::EnemyReadied { enemy: eid });
+            newly_readied.push(eid);
+        }
+    }
+    // RR p.10: "if an exhausted enemy at the same location as an investigator
+    // becomes ready, it engages as soon as it is readied." Runs after the
+    // (simultaneous, RR p.25) readying pass. Only newly-readied enemies are
+    // checked ("becomes ready"), and only those still unengaged —
+    // reengage_at_location's precondition is engaged_with == None, so an enemy
+    // that readied while still engaged keeps its existing engagement.
+    // newly_readied is in ascending EnemyId order (BTreeMap key order).
+    for eid in newly_readied {
+        if state.enemies[&eid].engaged_with.is_none() {
+            super::hunters::reengage_at_location(state, events, eid);
+        }
+    }
+}
+
+/// 4.5 Each investigator checks hand size.
+fn check_hand_size(_state: &mut crate::state::GameState, _events: &mut Vec<Event>) {
+    // TODO(#111): in player order, each investigator with more than 8
+    //   cards in hand discards down to 8 (Rules Reference p.25 step 4.5).
+    //   Needs an AwaitingInput producer for the discard choice. The call
+    //   site exists so the rule step is grep-able and #111 plugs in here
+    //   without changing the driver shape.
+}
+
+/// 4.2 Reset actions. Rules Reference p.25: "Flip each investigator's
+/// mini card back to its colored side. This indicates that the
+/// investigator's actions have been reset for his or her next turn."
+///
+/// The canonical action-refresh site. Sets `actions_remaining` to
+/// `ACTIONS_PER_TURN` for each Active investigator and emits
+/// `ActionsRemainingChanged` when the value changes. `rotate_to_active`
+/// no longer refreshes (step 2.2 is just "the turn begins");
+/// `start_scenario` seeds round 1. Eliminated investigators are skipped
+/// (Rules Reference p.10).
+fn reset_actions(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    for id in super::cursor::active_investigators_in_turn_order(state) {
+        let inv = state
+            .investigators
+            .get_mut(&id)
+            .expect("id from active_investigators_in_turn_order");
+        if inv.actions_remaining != ACTIONS_PER_TURN {
+            inv.actions_remaining = ACTIONS_PER_TURN;
+            events.push(Event::ActionsRemainingChanged {
+                investigator: id,
+                new_count: ACTIONS_PER_TURN,
+            });
+        }
+    }
+}
+
+/// 4.4 Each investigator draws 1 card and gains 1 resource. Rules
+/// Reference p.25: "In player order, each investigator draws 1 card.
+/// Once those cards have been drawn, each investigator gains 1
+/// resource." Two passes to honor that ordering: all draws first, then
+/// all resource gains.
+fn upkeep_draw_and_resource(state: &mut crate::state::GameState, events: &mut Vec<Event>) {
+    let ids = super::cursor::active_investigators_in_turn_order(state);
+    for &id in &ids {
+        super::cards::draw_one_with_deckout(state, events, id);
+    }
+    for &id in &ids {
+        super::cards::grant_resources(state, events, id, 1);
+    }
+}
+
+#[cfg(test)]
+mod investigation_phase_tests {
+    use super::*;
+    use crate::action::PlayerAction;
+    use crate::engine::dispatch::apply_player_action;
+    use crate::engine::outcome::EngineOutcome;
+    use crate::state::{InvestigatorId, Phase, Status, WindowKind};
+    use crate::test_support::{test_investigator, TestGame};
+
+    #[test]
+    fn mulligan_completion_kicks_off_investigation_phase() {
+        // After the last investigator mulligans, setup ends and the
+        // Investigation phase begins (Rules Reference p.27: no action
+        // windows during setup; the game begins after mulligans).
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+        state.active_investigator = None;
+        state.mulligan_pending = Some(InvestigatorId(1));
+
+        let mut events = Vec::new();
+        let outcome = apply_player_action(
+            &mut state,
+            &mut events,
+            &PlayerAction::Mulligan {
+                investigator: InvestigatorId(1),
+                indices_to_redraw: vec![],
+            },
+        );
+
+        assert!(matches!(outcome, EngineOutcome::Done));
+        assert_eq!(
+            state.mulligan_pending, None,
+            "mulligan cursor clears once every investigator has mulliganed"
+        );
+        assert_eq!(
+            state.active_investigator,
+            Some(InvestigatorId(1)),
+            "Investigation phase kicks off and rotates to the lead after mulligan completes"
+        );
+        // PhaseStarted(Investigation) fires at mulligan completion (not
+        // during StartScenario) AND precedes the post-2.1 window — the
+        // printed 2.1 → window order.
+        let phase_started = events.iter().position(|e| {
+            matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Investigation
+                }
+            )
+        });
+        let window_opened = events.iter().position(|e| {
+            matches!(
+                e,
+                Event::WindowOpened {
+                    kind: WindowKind::InvestigationBegins
+                }
+            )
+        });
+        let phase_started = phase_started.expect("PhaseStarted(Investigation) must fire");
+        let window_opened =
+            window_opened.expect("WindowOpened(InvestigationBegins) must fire at phase start");
+        assert!(
+            phase_started < window_opened,
+            "PhaseStarted (2.1) must precede the post-2.1 InvestigationBegins window"
+        );
+    }
+
+    #[test]
+    fn investigation_phase_emits_phase_started_and_rotates_to_lead() {
+        // Two investigators; investigation_phase should emit
+        // PhaseStarted(Investigation), open the post-2.1 InvestigationBegins
+        // window (which auto-skips in tests — no card registry installed),
+        // and then rotate to the first investigator in turn_order
+        // (Rules Reference p.24 step 2.1 → window → step 2.2 lead-first).
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2)];
+        state.active_investigator = None;
+
+        let mut events = Vec::new();
+        investigation_phase(&mut state, &mut events);
+
+        assert_eq!(
+            state.active_investigator,
+            Some(InvestigatorId(1)),
+            "investigation_phase must rotate to the lead (first in turn_order)"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Investigation
+                }
+            )),
+            "PhaseStarted(Investigation) must be emitted"
+        );
+        assert!(
+            !events
+                .iter()
+                .any(|e| matches!(e, Event::ActionsRemainingChanged { .. })),
+            "rotate no longer emits ActionsRemainingChanged (actions reset at Upkeep 4.2)"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::WindowOpened {
+                    kind: WindowKind::InvestigationBegins
+                }
+            )),
+            "investigation_phase opens the post-2.1 InvestigationBegins window"
+        );
+    }
+
+    #[test]
+    fn investigation_phase_with_empty_turn_order_parks() {
+        // Degenerate (cannot occur in real gameplay): no investigators.
+        // The InvestigationBegins continuation finds no active
+        // investigator and PARKS — active stays None, no PhaseEnded, no
+        // advance. Locks in the cascade-breaker behavior (see spec
+        // "All-eliminated / no-active-investigator handling").
+        //
+        // Phase starts as Investigation (matching the real call-site
+        // shape: step_phase sets state.phase before calling
+        // investigation_phase).
+        let mut state = TestGame::default().with_phase(Phase::Investigation).build();
+        state.turn_order.clear();
+        state.active_investigator = None;
+
+        let mut events = Vec::new();
+        investigation_phase(&mut state, &mut events);
+
+        assert_eq!(
+            state.active_investigator, None,
+            "no investigator to rotate to"
+        );
+        assert_eq!(state.phase, Phase::Investigation, "phase must not advance");
+        assert!(
+            !events.iter().any(|e| matches!(
+                e,
+                Event::PhaseEnded {
+                    phase: Phase::Investigation
+                }
+            )),
+            "parking must NOT end the phase (auto-advancing would loop the round)"
+        );
+    }
+
+    #[test]
+    fn investigation_phase_skips_defeated_lead_and_picks_first_active() {
+        // Investigator 1 (lead) is Killed; investigator 2 is Active.
+        // investigation_phase must skip Id(1) and rotate to Id(2).
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2)];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(1))
+            .unwrap()
+            .status = Status::Killed;
+        state.active_investigator = None;
+
+        let mut events = Vec::new();
+        investigation_phase(&mut state, &mut events);
+
+        assert_eq!(
+            state.active_investigator,
+            Some(InvestigatorId(2)),
+            "investigation_phase must skip the Killed lead and rotate to the first Active investigator"
+        );
+    }
+
+    #[test]
+    fn end_turn_for_last_investigator_ends_phase_and_steps_to_enemy() {
+        // Single investigator ends their turn: TurnEnded (2.2.2), then
+        // PhaseEnded(Investigation) (2.3) from investigation_phase_end,
+        // then the cascade enters the Enemy phase.
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Investigation)
+            .with_active_investigator(InvestigatorId(1))
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+
+        let mut events = Vec::new();
+        let outcome = end_turn(&mut state, &mut events);
+
+        assert!(matches!(outcome, EngineOutcome::Done));
+        assert!(
+            events.iter().any(|e| matches!(e, Event::TurnEnded { investigator } if *investigator == InvestigatorId(1))),
+            "step 2.2.2 emits TurnEnded"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::PhaseEnded {
+                    phase: Phase::Investigation
+                }
+            )),
+            "step 2.3 emits PhaseEnded(Investigation) via investigation_phase_end"
+        );
+        assert_eq!(
+            events
+                .iter()
+                .filter(|e| matches!(
+                    e,
+                    Event::PhaseEnded {
+                        phase: Phase::Investigation
+                    }
+                ))
+                .count(),
+            1,
+            "exactly one PhaseEnded(Investigation) — step_phase must not also emit it"
+        );
+        assert_ne!(
+            state.phase,
+            Phase::Investigation,
+            "phase advanced past Investigation"
+        );
+    }
+
+    #[test]
+    fn end_turn_rotates_to_next_active_and_opens_turn_window() {
+        // Two investigators: ending #1's turn returns to 2.2 for #2 and
+        // opens the InvestigatorTurnBegins window for them.
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_phase(Phase::Investigation)
+            .with_active_investigator(InvestigatorId(1))
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2)];
+
+        let mut events = Vec::new();
+        let outcome = end_turn(&mut state, &mut events);
+
+        assert!(matches!(outcome, EngineOutcome::Done));
+        assert_eq!(
+            state.active_investigator,
+            Some(InvestigatorId(2)),
+            "rotates to the next active investigator (return to 2.2)"
+        );
+        assert_eq!(
+            state.phase,
+            Phase::Investigation,
+            "phase does not end mid-round"
+        );
+        assert!(
+            !events.iter().any(|e| matches!(
+                e,
+                Event::PhaseEnded {
+                    phase: Phase::Investigation
+                }
+            )),
+            "phase must not end while an investigator is still to take a turn"
+        );
+    }
+
+    #[test]
+    fn step_phase_emits_no_phase_ended() {
+        // step_phase no longer emits PhaseEnded for any phase — each
+        // phase's *_end helper owns it. Direct Investigation→Enemy step:
+        // step_phase must NOT emit PhaseEnded(Investigation); the
+        // downstream cascade may emit PhaseEnded for Enemy/Upkeep via
+        // their own *_end helpers, but that's correct and expected.
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+
+        let mut events = Vec::new();
+        step_phase(&mut state, &mut events);
+
+        assert_eq!(
+            events
+                .iter()
+                .filter(|e| matches!(e, Event::PhaseEnded { phase: Phase::Investigation }))
+                .count(),
+            0,
+            "step_phase must emit no PhaseEnded(Investigation) — investigation_phase_end owns it. events = {events:?}"
+        );
+    }
+
+    #[test]
+    fn investigation_entry_emits_phase_started_then_windows_then_lead_active() {
+        // Round ≥2 entry via step_phase (Mythos→Investigation) auto-skips
+        // both windows (no registry → nothing Fast-eligible) and lands
+        // the lead active, with no PhaseEnded yet.
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+        state.active_investigator = None;
+
+        let mut events = Vec::new();
+        step_phase(&mut state, &mut events); // Mythos→Investigation
+
+        assert_eq!(state.phase, Phase::Investigation);
+        assert_eq!(state.active_investigator, Some(InvestigatorId(1)));
+        assert!(events.iter().any(|e| matches!(
+            e,
+            Event::PhaseStarted {
+                phase: Phase::Investigation
+            }
+        )));
+        assert!(events.iter().any(|e| matches!(
+            e,
+            Event::WindowOpened {
+                kind: WindowKind::InvestigationBegins
+            }
+        )));
+        assert!(events.iter().any(|e| matches!(
+            e,
+            Event::WindowOpened {
+                kind: WindowKind::InvestigatorTurnBegins
+            }
+        )));
+        assert!(!events.iter().any(|e| matches!(
+            e,
+            Event::PhaseEnded {
+                phase: Phase::Investigation
+            }
+        )));
+    }
+}
+
+#[cfg(test)]
+mod mythos_phase_tests {
+    use super::*;
+    use crate::state::{InvestigatorId, Phase, Status};
+    use crate::test_support::{test_investigator, TestGame};
+
+    #[test]
+    fn mythos_phase_emits_phase_started_and_seeds_draw_pending() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2)];
+        state.mythos_draw_pending = None;
+        let mut events = Vec::new();
+
+        mythos_phase(&mut state, &mut events);
+
+        assert_eq!(state.mythos_draw_pending, Some(InvestigatorId(1)));
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Mythos
+                }
+            )),
+            "must emit PhaseStarted(Mythos); events = {events:?}"
+        );
+    }
+
+    #[test]
+    fn mythos_phase_with_empty_turn_order_opens_after_draws_window_inline() {
+        let mut state = TestGame::default().with_phase(Phase::Mythos).build();
+        state.turn_order.clear();
+        state.mythos_draw_pending = None;
+        let mut events = Vec::new();
+
+        mythos_phase(&mut state, &mut events);
+
+        // No drawers → open_fast_window runs for MythosAfterDraws,
+        // which auto-skips (no Fast eligibility), runs continuation
+        // (mythos_phase_end), which steps into Investigation.
+        assert_eq!(state.mythos_draw_pending, None);
+        assert_eq!(state.phase, Phase::Investigation);
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::WindowOpened {
+                    kind: WindowKind::MythosAfterDraws
+                }
+            )),
+            "must emit WindowOpened(MythosAfterDraws); events = {events:?}"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::WindowClosed {
+                    kind: WindowKind::MythosAfterDraws
+                }
+            )),
+            "must emit WindowClosed(MythosAfterDraws); events = {events:?}"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::PhaseEnded {
+                    phase: Phase::Mythos
+                }
+            )),
+            "must emit PhaseEnded(Mythos); events = {events:?}"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Investigation
+                }
+            )),
+            "must emit PhaseStarted(Investigation); events = {events:?}"
+        );
+    }
+
+    #[test]
+    fn mythos_phase_end_emits_phase_ended_and_steps_to_investigation() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+        let mut events = Vec::new();
+
+        mythos_phase_end(&mut state, &mut events);
+
+        assert_eq!(state.phase, Phase::Investigation);
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::PhaseEnded {
+                    phase: Phase::Mythos
+                }
+            )),
+            "must emit PhaseEnded(Mythos); events = {events:?}"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Investigation
+                }
+            )),
+            "must emit PhaseStarted(Investigation); events = {events:?}"
+        );
+    }
+
+    /// Site 1 fix (Rules Reference p.10): when the lead investigator in
+    /// `turn_order` is eliminated, `mythos_phase` must seed
+    /// `mythos_draw_pending` with the first Active investigator rather
+    /// than blindly taking `turn_order.first()`.
+    #[test]
+    fn mythos_phase_skips_eliminated_lead_when_seeding_cursor() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2)];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(1))
+            .unwrap()
+            .status = Status::Killed;
+        state.mythos_draw_pending = None;
+        let mut events = Vec::new();
+
+        mythos_phase(&mut state, &mut events);
+
+        assert_eq!(
+            state.mythos_draw_pending,
+            Some(InvestigatorId(2)),
+            "cursor must point to the first Active investigator, not the Killed lead"
+        );
+    }
+
+    /// All investigators in `turn_order` are eliminated. `mythos_phase`
+    /// must treat this the same as an empty `turn_order`: seed to None
+    /// and open `MythosAfterDraws` inline, which auto-skips and drives
+    /// `mythos_phase_end`, transitioning to Investigation.
+    ///
+    /// This is the non-empty-`turn_order` analogue of
+    /// `mythos_phase_with_empty_turn_order_opens_after_draws_window_inline`.
+    #[test]
+    fn mythos_phase_with_all_investigators_eliminated_opens_after_draws_window() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(1))
+            .unwrap()
+            .status = Status::Killed;
+        state.mythos_draw_pending = None;
+        let mut events = Vec::new();
+
+        mythos_phase(&mut state, &mut events);
+
+        assert_eq!(state.mythos_draw_pending, None);
+        assert_eq!(
+            state.phase,
+            Phase::Investigation,
+            "no Active drawers → MythosAfterDraws fires inline → Investigation"
+        );
+    }
+
+    /// Site 2 fix (Rules Reference p.10): when advancing the cursor
+    /// after a completed draw, eliminated investigators in the middle of
+    /// `turn_order` must be skipped. Here inv2 is Killed; the cursor must
+    /// advance from inv1 to inv3.
+    #[test]
+    fn advance_mythos_draw_pending_skips_eliminated_middle_investigator() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_investigator(test_investigator(3))
+            .with_phase(Phase::Mythos)
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2), InvestigatorId(3)];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(2))
+            .unwrap()
+            .status = Status::Killed;
+        // Simulate: inv1 has just completed their draw chain.
+        state.mythos_draw_pending = Some(InvestigatorId(1));
+        let mut events = Vec::new();
+
+        super::super::encounter::advance_mythos_draw_pending(&mut state, &mut events);
+
+        assert_eq!(
+            state.mythos_draw_pending,
+            Some(InvestigatorId(3)),
+            "cursor must skip the Killed inv2 and land on Active inv3"
+        );
+    }
+
+    #[test]
+    fn first_active_investigator_finds_first_active_skipping_eliminated() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_investigator(test_investigator(3))
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2), InvestigatorId(3)];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(1))
+            .unwrap()
+            .status = Status::Killed;
+        state
+            .investigators
+            .get_mut(&InvestigatorId(2))
+            .unwrap()
+            .status = Status::Insane;
+
+        assert_eq!(
+            super::super::cursor::first_active_investigator(&state),
+            Some(InvestigatorId(3)),
+            "first Active in turn_order after skipping eliminated"
+        );
+    }
+
+    #[test]
+    fn first_active_investigator_returns_none_when_all_eliminated() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(1))
+            .unwrap()
+            .status = Status::Killed;
+
+        assert_eq!(
+            super::super::cursor::first_active_investigator(&state),
+            None
+        );
+    }
+
+    #[test]
+    fn first_active_investigator_returns_none_when_turn_order_empty() {
+        let state = TestGame::default().build();
+        assert_eq!(
+            super::super::cursor::first_active_investigator(&state),
+            None
+        );
+    }
+
+    #[test]
+    fn next_active_investigator_after_skips_eliminated_middle() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_investigator(test_investigator(3))
+            .with_investigator(test_investigator(4))
+            .build();
+        state.turn_order = vec![
+            InvestigatorId(1),
+            InvestigatorId(2),
+            InvestigatorId(3),
+            InvestigatorId(4),
+        ];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(2))
+            .unwrap()
+            .status = Status::Killed;
+
+        assert_eq!(
+            super::super::cursor::next_active_investigator_after(&state, InvestigatorId(1)),
+            Some(InvestigatorId(3)),
+            "advance from 1 skips Killed 2, lands on 3"
+        );
+        assert_eq!(
+            super::super::cursor::next_active_investigator_after(&state, InvestigatorId(3)),
+            Some(InvestigatorId(4)),
+            "advance from 3 lands on 4"
+        );
+        assert_eq!(
+            super::super::cursor::next_active_investigator_after(&state, InvestigatorId(4)),
+            None,
+            "advance past the last entry returns None"
+        );
+    }
+
+    #[test]
+    fn next_active_investigator_after_returns_none_when_current_not_in_turn_order() {
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+
+        assert_eq!(
+            super::super::cursor::next_active_investigator_after(&state, InvestigatorId(99)),
+            None
+        );
+    }
+
+    #[test]
+    fn next_active_investigator_after_works_when_current_is_non_active() {
+        // Defeated-mid-loop semantics: `current` may be Killed by the
+        // time we advance from them. The cursor still finds the right
+        // successor.
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .build();
+        state.turn_order = vec![InvestigatorId(1), InvestigatorId(2)];
+        state
+            .investigators
+            .get_mut(&InvestigatorId(1))
+            .unwrap()
+            .status = Status::Killed;
+
+        assert_eq!(
+            super::super::cursor::next_active_investigator_after(&state, InvestigatorId(1)),
+            Some(InvestigatorId(2)),
+            "current=1 is non-Active but turn_order still anchors the index"
+        );
+    }
+}
+
+#[cfg(test)]
+mod upkeep_phase_tests {
+    use super::*;
+    use crate::action::{Action, PlayerAction};
+    use crate::engine::{apply, EngineOutcome};
+    use crate::event::Event;
+    use crate::state::{
+        CardCode, CardInPlay, CardInstanceId, EnemyId, InvestigatorId, LocationId, Phase, Status,
+    };
+    use crate::test_support::{test_enemy, test_investigator, test_location, TestGame};
+    use crate::{assert_event, assert_event_sequence, assert_no_event};
+
+    #[test]
+    fn upkeep_phase_emits_phase_started_and_auto_skips_to_mythos() {
+        // No Fast-eligible cards / no reactions installed → the post-4.1
+        // window auto-skips inline, the continuation runs, and the
+        // cascade lands in Mythos.
+        let id = InvestigatorId(1);
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Enemy)
+            .build();
+        state.turn_order = vec![id];
+        state.active_investigator = None;
+
+        let mut events = Vec::new();
+        step_phase(&mut state, &mut events); // Enemy → Upkeep, cascades to Mythos
+
+        let pos = |pred: &dyn Fn(&Event) -> bool| events.iter().position(pred);
+        let started = pos(&|e| {
+            matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Upkeep
+                }
+            )
+        })
+        .expect("PhaseStarted(Upkeep)");
+        let w_open = pos(&|e| {
+            matches!(
+                e,
+                Event::WindowOpened {
+                    kind: WindowKind::UpkeepBegins
+                }
+            )
+        })
+        .expect("WindowOpened(UpkeepBegins)");
+        let w_close = pos(&|e| {
+            matches!(
+                e,
+                Event::WindowClosed {
+                    kind: WindowKind::UpkeepBegins
+                }
+            )
+        })
+        .expect("WindowClosed(UpkeepBegins)");
+        let ended = pos(&|e| {
+            matches!(
+                e,
+                Event::PhaseEnded {
+                    phase: Phase::Upkeep
+                }
+            )
+        })
+        .expect("PhaseEnded(Upkeep)");
+        let mythos = pos(&|e| {
+            matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Mythos
+                }
+            )
+        })
+        .expect("PhaseStarted(Mythos)");
+        assert!(
+            started < w_open && w_open < w_close && w_close < ended && ended < mythos,
+            "upkeep sub-step events must be ordered 4.1 → window → 4.6 → Mythos 1.1; \
+             events = {events:?}"
+        );
+        assert_eq!(state.phase, Phase::Mythos, "cascade lands in Mythos");
+        assert!(
+            state.open_windows.is_empty(),
+            "UpkeepBegins must not persist on the stack"
+        );
+    }
+
+    #[test]
+    fn ready_exhausted_cards_readies_investigator_cards_and_enemies() {
+        let inv_id = InvestigatorId(1);
+        let enemy_id = EnemyId(1);
+        let mut inv = test_investigator(1);
+        let mut card = CardInPlay::enter_play(CardCode("01000".into()), CardInstanceId(1));
+        card.exhausted = true;
+        inv.cards_in_play = vec![card];
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.exhausted = true;
+        let mut state = TestGame::default()
+            .with_investigator(inv)
+            .with_enemy(enemy)
+            .build();
+        let mut events = Vec::new();
+
+        ready_exhausted_cards(&mut state, &mut events);
+
+        assert!(
+            !state.investigators[&inv_id].cards_in_play[0].exhausted,
+            "card readied"
+        );
+        assert!(!state.enemies[&enemy_id].exhausted, "enemy readied");
+        assert!(events.iter().any(|e| matches!(
+            e, Event::CardReadied { investigator, instance_id, .. }
+            if *investigator == inv_id && *instance_id == CardInstanceId(1))));
+        assert!(events.iter().any(|e| matches!(
+            e, Event::EnemyReadied { enemy } if *enemy == enemy_id)));
+    }
+
+    #[test]
+    fn ready_exhausted_cards_reengages_co_located_unengaged_enemy() {
+        let inv_id = InvestigatorId(1);
+        let enemy_id = EnemyId(1);
+        let loc = test_location(10, "Synth Loc");
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.exhausted = true; // exhausted + disengaged, e.g. survived a successful Evade
+        enemy.current_location = Some(LocationId(10));
+        let mut state = TestGame::default()
+            .with_investigator_at(test_investigator(1), LocationId(10))
+            .with_location(loc)
+            .with_enemy(enemy)
+            .with_turn_order([inv_id])
+            .build();
+        let mut events = Vec::new();
+
+        ready_exhausted_cards(&mut state, &mut events);
+
+        assert!(!state.enemies[&enemy_id].exhausted, "enemy readied");
+        assert_eq!(
+            state.enemies[&enemy_id].engaged_with,
+            Some(inv_id),
+            "readied enemy re-engages the co-located investigator (RR p.10)"
+        );
+        assert_event!(events, Event::EnemyReadied { enemy } if *enemy == enemy_id);
+        assert_event!(events, Event::EnemyEngaged { investigator, .. } if *investigator == inv_id);
+        assert_event_sequence!(
+            events,
+            Event::EnemyReadied { .. },
+            Event::EnemyEngaged { .. },
+        );
+    }
+
+    #[test]
+    fn ready_exhausted_cards_leaves_ready_cards_untouched() {
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.exhausted = false; // already ready
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_enemy(enemy)
+            .build();
+        let mut events = Vec::new();
+
+        ready_exhausted_cards(&mut state, &mut events);
+
+        assert!(
+            events.is_empty(),
+            "no readying events for already-ready cards"
+        );
+    }
+
+    #[test]
+    fn ready_exhausted_cards_no_engage_when_no_co_located_investigator() {
+        let enemy_id = EnemyId(1);
+        let inv_id = InvestigatorId(1);
+        let loc = test_location(10, "Synth Loc");
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.exhausted = true;
+        enemy.current_location = Some(LocationId(10));
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1)) // current_location stays None — NOT co-located
+            .with_location(loc)
+            .with_enemy(enemy)
+            .with_turn_order([inv_id])
+            .build();
+        let mut events = Vec::new();
+
+        ready_exhausted_cards(&mut state, &mut events);
+
+        assert!(!state.enemies[&enemy_id].exhausted, "enemy readied");
+        assert_eq!(
+            state.enemies[&enemy_id].engaged_with, None,
+            "no investigator at the enemy's location → no engagement"
+        );
+        assert_no_event!(events, Event::EnemyEngaged { .. });
+    }
+
+    #[test]
+    fn ready_exhausted_cards_keeps_existing_engagement_no_duplicate() {
+        let enemy_id = EnemyId(1);
+        let inv_id = InvestigatorId(1);
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.exhausted = true; // exhausted but still engaged (e.g. attacked last Enemy phase)
+        enemy.engaged_with = Some(inv_id);
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_enemy(enemy)
+            .build();
+        let mut events = Vec::new();
+
+        ready_exhausted_cards(&mut state, &mut events);
+
+        assert!(!state.enemies[&enemy_id].exhausted, "enemy readied");
+        assert_eq!(
+            state.enemies[&enemy_id].engaged_with,
+            Some(inv_id),
+            "an already-engaged enemy keeps its engagement"
+        );
+        assert_no_event!(events, Event::EnemyEngaged { .. });
+    }
+
+    #[test]
+    fn upkeep_draw_and_resource_draws_and_grants_per_active_investigator() {
+        let (a, b, c) = (InvestigatorId(1), InvestigatorId(2), InvestigatorId(3));
+        let mut inv_a = test_investigator(1);
+        inv_a.deck = vec![CardCode::new("01000")];
+        let mut inv_b = test_investigator(2);
+        inv_b.deck = vec![CardCode::new("01001")];
+        let mut inv_c = test_investigator(3);
+        inv_c.status = Status::Resigned; // eliminated → skipped
+        inv_c.deck = vec![CardCode::new("01002")];
+        let res_a = inv_a.resources;
+        let res_b = inv_b.resources;
+        let res_c = inv_c.resources;
+        let hand_a = inv_a.hand.len();
+        let mut state = TestGame::default()
+            .with_investigator(inv_a)
+            .with_investigator(inv_b)
+            .with_investigator(inv_c)
+            .build();
+        state.turn_order = vec![a, b, c];
+        let mut events = Vec::new();
+
+        upkeep_draw_and_resource(&mut state, &mut events);
+
+        assert_eq!(state.investigators[&a].resources, res_a + 1);
+        assert_eq!(state.investigators[&b].resources, res_b + 1);
+        assert_eq!(
+            state.investigators[&c].resources, res_c,
+            "eliminated investigator skipped"
+        );
+        assert_eq!(state.investigators[&a].hand.len(), hand_a + 1);
+        assert_eq!(
+            state.investigators[&c].deck.len(),
+            1,
+            "eliminated investigator did not draw"
+        );
+    }
+
+    #[test]
+    fn upkeep_draw_and_resource_two_pass_ordering() {
+        // All CardsDrawn events precede all ResourcesGained events.
+        let (a, b) = (InvestigatorId(1), InvestigatorId(2));
+        let mut inv_a = test_investigator(1);
+        inv_a.deck = vec![CardCode::new("01000")];
+        let mut inv_b = test_investigator(2);
+        inv_b.deck = vec![CardCode::new("01001")];
+        let mut state = TestGame::default()
+            .with_investigator(inv_a)
+            .with_investigator(inv_b)
+            .build();
+        state.turn_order = vec![a, b];
+        let mut events = Vec::new();
+
+        upkeep_draw_and_resource(&mut state, &mut events);
+
+        let last_draw = events
+            .iter()
+            .rposition(|e| matches!(e, Event::CardsDrawn { .. }))
+            .expect("draws");
+        let first_gain = events
+            .iter()
+            .position(|e| matches!(e, Event::ResourcesGained { .. }))
+            .expect("gains");
+        assert!(
+            last_draw < first_gain,
+            "all draws must precede all resource gains"
+        );
+    }
+
+    #[test]
+    fn reset_actions_sets_active_to_per_turn_and_skips_eliminated() {
+        let (a, b) = (InvestigatorId(1), InvestigatorId(2));
+        let mut inv_a = test_investigator(1);
+        inv_a.actions_remaining = 0;
+        let mut inv_b = test_investigator(2);
+        inv_b.actions_remaining = 0;
+        inv_b.status = Status::Killed;
+        let mut state = TestGame::default()
+            .with_investigator(inv_a)
+            .with_investigator(inv_b)
+            .build();
+        state.turn_order = vec![a, b];
+        let mut events = Vec::new();
+
+        reset_actions(&mut state, &mut events);
+
+        assert_eq!(state.investigators[&a].actions_remaining, ACTIONS_PER_TURN);
+        assert_eq!(
+            state.investigators[&b].actions_remaining, 0,
+            "eliminated skipped"
+        );
+        assert!(events.iter().any(|e| matches!(
+            e, Event::ActionsRemainingChanged { investigator, new_count }
+            if *investigator == a && *new_count == ACTIONS_PER_TURN)));
+        assert!(!events.iter().any(|e| matches!(
+            e, Event::ActionsRemainingChanged { investigator, .. } if *investigator == b)));
+    }
+
+    #[test]
+    fn reset_actions_emits_nothing_for_already_full() {
+        // Emit-on-change semantics: when actions_remaining already equals
+        // ACTIONS_PER_TURN, reset_actions makes no state change and emits
+        // no ActionsRemainingChanged event.
+        let id = InvestigatorId(1);
+        let mut inv = test_investigator(1);
+        inv.actions_remaining = ACTIONS_PER_TURN;
+        let mut state = TestGame::default().with_investigator(inv).build();
+        state.turn_order = vec![id];
+        let mut events = Vec::new();
+
+        reset_actions(&mut state, &mut events);
+
+        assert_eq!(state.investigators[&id].actions_remaining, ACTIONS_PER_TURN);
+        assert!(events.is_empty(), "no event when value is unchanged");
+    }
+
+    #[test]
+    fn rotate_to_active_does_not_refresh_actions() {
+        let id = InvestigatorId(1);
+        let mut inv = test_investigator(1);
+        inv.actions_remaining = 1;
+        let mut state = TestGame::default().with_investigator(inv).build();
+        let mut events = Vec::new();
+
+        rotate_to_active(&mut state, &mut events, id);
+
+        assert_eq!(state.active_investigator, Some(id));
+        assert_eq!(
+            state.investigators[&id].actions_remaining, 1,
+            "rotate must not refresh actions"
+        );
+        assert!(
+            events.is_empty(),
+            "rotate no longer emits ActionsRemainingChanged"
+        );
+    }
+
+    #[test]
+    fn round_increments_on_mythos_entry_via_driver() {
+        // After the Upkeep→Mythos cascade, state.round bumps by 1.
+        // The bump now lives in mythos_phase step 1.1 (this task);
+        // the test asserts observable behavior, which is unchanged.
+        let id = InvestigatorId(1);
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Upkeep)
+            .build();
+        state.turn_order = vec![id];
+        state.active_investigator = None;
+        state.round = 4;
+
+        let mut events = Vec::new();
+        step_phase(&mut state, &mut events); // Upkeep → ... → Mythos via the cascade
+
+        assert_eq!(state.round, 5, "round bumps on Mythos entry");
+        assert_eq!(state.phase, Phase::Mythos);
+    }
+
+    #[test]
+    fn end_turn_cascades_through_upkeep_to_mythos_draw_pending() {
+        // Single investigator, non-empty deck, an exhausted in-play card.
+        // After EndTurn: card readied, hand +1, resources +1, landed in
+        // Mythos with draw pending and round bumped.
+        let id = InvestigatorId(1);
+        let mut inv = test_investigator(1);
+        inv.actions_remaining = 0;
+        inv.deck = vec![CardCode::new("01000"), CardCode::new("01001")];
+        let mut card = CardInPlay::enter_play(CardCode::new("01002"), CardInstanceId(1));
+        card.exhausted = true;
+        inv.cards_in_play = vec![card];
+        let res_before = inv.resources;
+        let hand_before = inv.hand.len();
+        let mut state = TestGame::default()
+            .with_investigator(inv)
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![id];
+        state.active_investigator = Some(id);
+        state.round = 1;
+
+        let result = apply(state, Action::Player(PlayerAction::EndTurn));
+
+        assert_eq!(result.outcome, EngineOutcome::Done);
+        assert_eq!(result.state.phase, Phase::Mythos);
+        assert_eq!(result.state.round, 2, "round bumped on Mythos entry");
+        assert_eq!(result.state.mythos_draw_pending, Some(id));
+        assert_eq!(result.state.active_investigator, None);
+        assert!(
+            !result.state.investigators[&id].cards_in_play[0].exhausted,
+            "readied"
+        );
+        assert_eq!(
+            result.state.investigators[&id].resources,
+            res_before + 1,
+            "gained 1"
+        );
+        assert_eq!(
+            result.state.investigators[&id].hand.len(),
+            hand_before + 1,
+            "drew 1"
+        );
+    }
+}
+
+#[cfg(test)]
+mod enemy_phase_tests {
+    use super::*;
+    use crate::action::{Action, InputResponse, PlayerAction};
+    use crate::assert_event;
+    use crate::engine::dispatch::resolve_input;
+    use crate::engine::{apply, EngineOutcome};
+    use crate::state::{
+        EnemyId, FastActorScope, InvestigatorId, LocationId, OpenWindow, Phase, Status, WindowKind,
+    };
+    use crate::test_support::{test_enemy, test_investigator, test_location, TestGame};
+
+    #[test]
+    fn enemy_phase_runs_hunters_then_attack_loop_when_no_tie() {
+        let mut loc_a = test_location(1, "A");
+        let mut loc_b = test_location(2, "B");
+        loc_a.connections = vec![LocationId(2)];
+        loc_b.connections = vec![LocationId(1)];
+        let mut inv = test_investigator(1);
+        inv.current_location = Some(LocationId(2));
+        let mut hunter = test_enemy(1, "Hunter");
+        hunter.hunter = true;
+        hunter.current_location = Some(LocationId(1));
+        let mut state = TestGame::new()
+            .with_phase(Phase::Investigation)
+            .with_location(loc_a)
+            .with_location(loc_b)
+            .with_investigator(inv)
+            .with_active_investigator(InvestigatorId(1))
+            .with_turn_order([InvestigatorId(1)])
+            .with_enemy(hunter)
+            .build();
+        let mut events = Vec::new();
+        let outcome = end_turn(&mut state, &mut events);
+        assert_eq!(outcome, EngineOutcome::Done);
+        // No registry installed → the attack window auto-skips inline and
+        // the cascade runs Enemy→Upkeep→Mythos within this same call (same
+        // as `enemy_phase_emits_phase_started_and_cascades_to_mythos...`).
+        // The hunter still moved + engaged during step 3.2, and the first
+        // attack window still opened — asserted via the event stream below.
+        assert_eq!(state.phase, Phase::Mythos);
+        assert_eq!(
+            state.enemies[&EnemyId(1)].current_location,
+            Some(LocationId(2))
+        );
+        assert_event!(events, Event::EnemyEngaged { enemy, .. } if *enemy == EnemyId(1));
+        assert_event!(events, Event::WindowOpened { kind } if *kind == WindowKind::BeforeInvestigatorAttacked);
+    }
+
+    #[test]
+    fn enemy_phase_suspends_on_hunter_tie_then_resumes_into_attack_loop() {
+        let mut loc_a = test_location(1, "A");
+        let mut loc_b = test_location(2, "B");
+        let mut loc_c = test_location(3, "C");
+        let mut loc_d = test_location(4, "D");
+        loc_a.connections = vec![LocationId(2), LocationId(3)];
+        loc_b.connections = vec![LocationId(1), LocationId(4)];
+        loc_c.connections = vec![LocationId(1), LocationId(4)];
+        loc_d.connections = vec![LocationId(2), LocationId(3)];
+        let mut inv = test_investigator(1);
+        inv.current_location = Some(LocationId(4));
+        let mut hunter = test_enemy(1, "Hunter");
+        hunter.hunter = true;
+        hunter.current_location = Some(LocationId(1));
+        let mut state = TestGame::new()
+            .with_phase(Phase::Investigation)
+            .with_location(loc_a)
+            .with_location(loc_b)
+            .with_location(loc_c)
+            .with_location(loc_d)
+            .with_investigator(inv)
+            .with_active_investigator(InvestigatorId(1))
+            .with_turn_order([InvestigatorId(1)])
+            .with_enemy(hunter)
+            .build();
+        let mut events = Vec::new();
+        let outcome = end_turn(&mut state, &mut events);
+        assert!(matches!(outcome, EngineOutcome::AwaitingInput { .. }));
+        assert_eq!(state.phase, Phase::Enemy);
+        let mut ev2 = Vec::new();
+        let resumed = resolve_input(
+            &mut state,
+            &mut ev2,
+            &InputResponse::PickLocation(LocationId(2)),
+        );
+        assert_eq!(resumed, EngineOutcome::Done);
+        assert_event!(ev2, Event::WindowOpened { kind } if *kind == WindowKind::BeforeInvestigatorAttacked);
+        // With no registry the attack window auto-skips and the cascade runs
+        // Enemy->Upkeep->Mythos within the same resume call (same as the no-tie test).
+        assert_eq!(state.phase, Phase::Mythos);
+    }
+
+    #[test]
+    fn resolve_attacks_for_investigator_fires_engaged_ready_enemy_and_exhausts() {
+        let inv_id = InvestigatorId(1);
+        let enemy_id = EnemyId(1);
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.engaged_with = Some(inv_id);
+        enemy.attack_damage = 1;
+        enemy.attack_horror = 0;
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_enemy(enemy)
+            .build();
+        let mut events = Vec::new();
+
+        super::super::combat::resolve_attacks_for_investigator(&mut state, &mut events, inv_id);
+
+        // Damage placed.
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::DamageTaken { investigator, amount: 1 } if *investigator == inv_id
+            )),
+            "expected DamageTaken {{ amount: 1 }}; events = {events:?}"
+        );
+
+        // Enemy exhausted in state and event.
+        assert!(
+            state.enemies[&enemy_id].exhausted,
+            "enemy must be exhausted"
+        );
+        assert!(
+            events.iter().any(|e| matches!(
+                e,
+                Event::EnemyExhausted { enemy } if *enemy == enemy_id
+            )),
+            "expected EnemyExhausted; events = {events:?}"
+        );
+
+        // Ordering: DamageTaken precedes EnemyExhausted (post-attack exhaust).
+        let damage_pos = events
+            .iter()
+            .position(|e| matches!(e, Event::DamageTaken { .. }))
+            .unwrap();
+        let exhaust_pos = events
+            .iter()
+            .position(|e| matches!(e, Event::EnemyExhausted { .. }))
+            .unwrap();
+        assert!(
+            damage_pos < exhaust_pos,
+            "DamageTaken must precede EnemyExhausted; events = {events:?}"
+        );
+    }
+
+    #[test]
+    fn resolve_attacks_for_investigator_excludes_exhausted_and_unengaged_enemies() {
+        let inv_id = InvestigatorId(1);
+
+        // Engaged but exhausted — must NOT attack.
+        let mut e1 = test_enemy(1, "Exhausted Engaged");
+        e1.engaged_with = Some(inv_id);
+        e1.exhausted = true;
+        e1.attack_damage = 5;
+
+        // Ready but unengaged — must NOT attack.
+        let mut e2 = test_enemy(2, "Ready Unengaged");
+        e2.engaged_with = None;
+        e2.attack_damage = 5;
+
+        // Ready engaged — the only one that attacks.
+        let mut e3 = test_enemy(3, "Ready Engaged");
+        e3.engaged_with = Some(inv_id);
+        e3.attack_damage = 1;
+
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_enemy(e1)
+            .with_enemy(e2)
+            .with_enemy(e3)
+            .build();
+        let mut events = Vec::new();
+
+        super::super::combat::resolve_attacks_for_investigator(&mut state, &mut events, inv_id);
+
+        // Exactly one DamageTaken (from e3, amount 1).
+        let damages: Vec<&Event> = events
+            .iter()
+            .filter(|e| matches!(e, Event::DamageTaken { .. }))
+            .collect();
+        assert_eq!(
+            damages.len(),
+            1,
+            "exactly one attacker should fire; events = {events:?}"
+        );
+        assert!(matches!(damages[0], Event::DamageTaken { amount: 1, .. }));
+
+        // Only e3 exhausted; e1 already was; e2 must remain ready.
+        assert!(
+            state.enemies[&EnemyId(1)].exhausted,
+            "e1 was already exhausted; still is"
+        );
+        assert!(
+            !state.enemies[&EnemyId(2)].exhausted,
+            "e2 must NOT exhaust (didn't attack)"
+        );
+        assert!(
+            state.enemies[&EnemyId(3)].exhausted,
+            "e3 attacked and exhausted"
+        );
+
+        // Exactly one EnemyExhausted event (e3). e1's prior-state exhausted doesn't re-emit.
+        let exhausted_events: Vec<&Event> = events
+            .iter()
+            .filter(|e| matches!(e, Event::EnemyExhausted { .. }))
+            .collect();
+        assert_eq!(exhausted_events.len(), 1);
+        assert!(matches!(
+            exhausted_events[0],
+            Event::EnemyExhausted { enemy: EnemyId(3) }
+        ));
+    }
+
+    #[test]
+    fn resolve_attacks_for_investigator_iterates_attackers_in_enemy_id_order() {
+        let inv_id = InvestigatorId(1);
+
+        let mut e_lower = test_enemy(2, "Lower id"); // EnemyId(2)
+        e_lower.engaged_with = Some(inv_id);
+        e_lower.attack_damage = 1;
+
+        let mut e_higher = test_enemy(10, "Higher id"); // EnemyId(10)
+        e_higher.engaged_with = Some(inv_id);
+        e_higher.attack_damage = 2;
+
+        let mut state = TestGame::default()
+            .with_investigator({
+                let mut inv = test_investigator(1);
+                inv.max_health = 100; // survive both attacks
+                inv
+            })
+            .with_enemy(e_higher) // insert in NON-id order to confirm BTreeMap ordering wins
+            .with_enemy(e_lower)
+            .build();
+        let mut events = Vec::new();
+
+        super::super::combat::resolve_attacks_for_investigator(&mut state, &mut events, inv_id);
+
+        // The two DamageTaken events must appear in EnemyId(2) → EnemyId(10) order
+        // (verifiable via their amounts: 1 then 2).
+        let damages: Vec<u8> = events
+            .iter()
+            .filter_map(|e| match e {
+                Event::DamageTaken { amount, .. } => Some(*amount),
+                _ => None,
+            })
+            .collect();
+        assert_eq!(
+            damages,
+            vec![1, 2],
+            "EnemyId order: 2 (dmg 1) before 10 (dmg 2)"
+        );
+    }
+
+    #[test]
+    fn resolve_attacks_for_investigator_early_breaks_when_target_defeated_mid_loop() {
+        let inv_id = InvestigatorId(1);
+
+        // EnemyId(1) deals the killing blow on its attack.
+        let mut e1 = test_enemy(1, "Killer");
+        e1.engaged_with = Some(inv_id);
+        e1.attack_damage = 1;
+
+        // EnemyId(2) must NOT attack (active check fails at loop top).
+        let mut e2 = test_enemy(2, "Bystander");
+        e2.engaged_with = Some(inv_id);
+        e2.attack_damage = 5;
+
+        let mut state = TestGame::default()
+            .with_investigator({
+                let mut inv = test_investigator(1);
+                inv.max_health = 1; // e1's attack defeats
+                inv
+            })
+            .with_enemy(e1)
+            .with_enemy(e2)
+            .build();
+        let mut events = Vec::new();
+
+        super::super::combat::resolve_attacks_for_investigator(&mut state, &mut events, inv_id);
+
+        // e1 attacked + exhausted.
+        assert!(
+            state.enemies[&EnemyId(1)].exhausted,
+            "e1 attacked, must exhaust"
+        );
+        // e2 did NOT attack and did NOT exhaust.
+        assert!(
+            !state.enemies[&EnemyId(2)].exhausted,
+            "e2 must not exhaust (early-break)"
+        );
+
+        let damages: Vec<&Event> = events
+            .iter()
+            .filter(|e| matches!(e, Event::DamageTaken { .. }))
+            .collect();
+        assert_eq!(
+            damages.len(),
+            1,
+            "only e1's attack lands; events = {events:?}"
+        );
+
+        let exhausted_events: Vec<&Event> = events
+            .iter()
+            .filter(|e| matches!(e, Event::EnemyExhausted { .. }))
+            .collect();
+        assert_eq!(exhausted_events.len(), 1);
+        assert!(matches!(
+            exhausted_events[0],
+            Event::EnemyExhausted { enemy: EnemyId(1) }
+        ));
+
+        // Investigator was defeated.
+        assert_eq!(state.investigators[&inv_id].status, Status::Killed);
+    }
+
+    #[test]
+    fn enemy_phase_emits_phase_started_and_cascades_to_mythos_in_no_eligibility_case() {
+        // 1 Active investigator, no engaged enemies. Auto-skip
+        // cascades through both windows + enemy_phase_end +
+        // Upkeep → Mythos.
+        let inv_id = InvestigatorId(1);
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![inv_id];
+        state.active_investigator = None;
+        let mut events = Vec::new();
+
+        step_phase(&mut state, &mut events); // Investigation → Enemy
+
+        // Positional ordering of the major events.
+        let pos = |pred: &dyn Fn(&Event) -> bool| events.iter().position(pred);
+        let started = pos(&|e| {
+            matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Enemy
+                }
+            )
+        })
+        .expect("PhaseStarted(Enemy)");
+        let w1_open = pos(&|e| {
+            matches!(
+                e,
+                Event::WindowOpened {
+                    kind: WindowKind::BeforeInvestigatorAttacked
+                }
+            )
+        })
+        .expect("WindowOpened(Before)");
+        let w1_close = pos(&|e| {
+            matches!(
+                e,
+                Event::WindowClosed {
+                    kind: WindowKind::BeforeInvestigatorAttacked
+                }
+            )
+        })
+        .expect("WindowClosed(Before)");
+        let w2_open = pos(&|e| {
+            matches!(
+                e,
+                Event::WindowOpened {
+                    kind: WindowKind::AfterAllInvestigatorsAttacked
+                }
+            )
+        })
+        .expect("WindowOpened(After)");
+        let w2_close = pos(&|e| {
+            matches!(
+                e,
+                Event::WindowClosed {
+                    kind: WindowKind::AfterAllInvestigatorsAttacked
+                }
+            )
+        })
+        .expect("WindowClosed(After)");
+        let ended = pos(&|e| {
+            matches!(
+                e,
+                Event::PhaseEnded {
+                    phase: Phase::Enemy
+                }
+            )
+        })
+        .expect("PhaseEnded(Enemy)");
+        let upkeep_started = pos(&|e| {
+            matches!(
+                e,
+                Event::PhaseStarted {
+                    phase: Phase::Upkeep
+                }
+            )
+        })
+        .expect("PhaseStarted(Upkeep)");
+
+        assert!(
+            started < w1_open
+                && w1_open < w1_close
+                && w1_close < w2_open
+                && w2_open < w2_close
+                && w2_close < ended
+                && ended < upkeep_started,
+            "ordered: 3.1 → BeforeInv window → AfterAll window → 3.4 → Upkeep 4.1; events = {events:?}"
+        );
+        assert_eq!(state.phase, Phase::Mythos, "cascade lands in Mythos");
+        assert_eq!(state.enemy_attack_pending, None, "cursor cleared at end");
+    }
+
+    #[test]
+    fn enemy_phase_with_two_investigators_iterates_in_turn_order() {
+        let id1 = InvestigatorId(1);
+        let id2 = InvestigatorId(2);
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![id1, id2];
+        state.active_investigator = None;
+        let mut events = Vec::new();
+
+        step_phase(&mut state, &mut events); // Investigation → Enemy
+
+        // Two BeforeInvestigatorAttacked windows + one AfterAll.
+        let before_opens: Vec<usize> = events
+            .iter()
+            .enumerate()
+            .filter_map(|(i, e)| {
+                matches!(
+                    e,
+                    Event::WindowOpened {
+                        kind: WindowKind::BeforeInvestigatorAttacked
+                    }
+                )
+                .then_some(i)
+            })
+            .collect();
+        let after_opens: Vec<usize> = events
+            .iter()
+            .enumerate()
+            .filter_map(|(i, e)| {
+                matches!(
+                    e,
+                    Event::WindowOpened {
+                        kind: WindowKind::AfterAllInvestigatorsAttacked
+                    }
+                )
+                .then_some(i)
+            })
+            .collect();
+        assert_eq!(before_opens.len(), 2, "one window per Active investigator");
+        assert_eq!(after_opens.len(), 1);
+        assert!(before_opens[0] < before_opens[1] && before_opens[1] < after_opens[0]);
+    }
+
+    #[test]
+    fn enemy_phase_skips_eliminated_investigator_in_advance() {
+        let id1 = InvestigatorId(1);
+        let id2 = InvestigatorId(2);
+        let id3 = InvestigatorId(3);
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_investigator(test_investigator(2))
+            .with_investigator(test_investigator(3))
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![id1, id2, id3];
+        state.active_investigator = None;
+        state.investigators.get_mut(&id2).unwrap().status = Status::Insane;
+        let mut events = Vec::new();
+
+        step_phase(&mut state, &mut events); // Investigation → Enemy
+
+        // Only 2 BeforeInvestigatorAttacked windows (id1 + id3).
+        let before_count = events
+            .iter()
+            .filter(|e| {
+                matches!(
+                    e,
+                    Event::WindowOpened {
+                        kind: WindowKind::BeforeInvestigatorAttacked
+                    }
+                )
+            })
+            .count();
+        assert_eq!(before_count, 2, "Insane id2 must be skipped");
+    }
+
+    #[test]
+    fn enemy_phase_with_all_eliminated_opens_after_all_directly() {
+        let id1 = InvestigatorId(1);
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![id1];
+        state.active_investigator = None;
+        state.investigators.get_mut(&id1).unwrap().status = Status::Killed;
+        let mut events = Vec::new();
+
+        step_phase(&mut state, &mut events); // Investigation → Enemy
+
+        // No BeforeInvestigatorAttacked windows — straight to AfterAll.
+        assert!(
+            events.iter().all(|e| !matches!(
+                e,
+                Event::WindowOpened {
+                    kind: WindowKind::BeforeInvestigatorAttacked
+                }
+            )),
+            "no per-investigator window when all are eliminated; events = {events:?}"
+        );
+        assert!(events.iter().any(|e| matches!(
+            e,
+            Event::WindowOpened {
+                kind: WindowKind::AfterAllInvestigatorsAttacked
+            }
+        )));
+        // With all investigators eliminated, the cascade keeps going:
+        // Enemy → Upkeep (no-op steps for empty Active set) → Mythos
+        // (mythos_draw_pending = None → auto-skip path) → Investigation.
+        // The point of this test is the structural shape — no
+        // BeforeInvestigatorAttacked window, AfterAll opens directly —
+        // not the terminal phase.
+        assert_eq!(state.phase, Phase::Investigation);
+    }
+
+    #[test]
+    fn enemy_phase_attack_lands_in_full_cascade() {
+        // 1 investigator engaged with 1 ready enemy. Full Investigation→Enemy→Upkeep→Mythos
+        // cascade; attack lands inside the BeforeInvestigatorAttacked continuation.
+        let inv_id = InvestigatorId(1);
+        let enemy_id = EnemyId(1);
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.engaged_with = Some(inv_id);
+        enemy.attack_damage = 1;
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_enemy(enemy)
+            .with_phase(Phase::Investigation)
+            .build();
+        state.turn_order = vec![inv_id];
+        state.active_investigator = None;
+        let mut events = Vec::new();
+
+        step_phase(&mut state, &mut events); // Investigation → Enemy
+
+        // The attack landed. Event-stream evidence — state.enemies's
+        // `exhausted` flag is reset by Upkeep step 4.3 later in the
+        // cascade (ready_exhausted_cards), so checking the post-cascade
+        // state directly would race the readying step. The
+        // DamageTaken + EnemyExhausted events emitted inside the
+        // BeforeInvestigatorAttacked continuation are the authoritative
+        // signal that the attack landed.
+        assert!(events.iter().any(|e| matches!(
+            e,
+            Event::DamageTaken { investigator, amount: 1 } if *investigator == inv_id
+        )));
+        assert!(events.iter().any(|e| matches!(
+            e,
+            Event::EnemyExhausted { enemy } if *enemy == enemy_id
+        )));
+
+        // Cascade landed in Mythos.
+        assert_eq!(state.phase, Phase::Mythos);
+    }
+
+    #[test]
+    fn step_phase_from_enemy_does_not_emit_phase_ended_enemy() {
+        // Direct unit-level check: step_phase emits no PhaseEnded itself,
+        // so the Enemy→Upkeep step must not emit PhaseEnded(Enemy)
+        // (enemy_phase_end owns that emit).
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_phase(Phase::Enemy)
+            .build();
+        state.turn_order = vec![InvestigatorId(1)];
+        state.active_investigator = None;
+        // Use a state where Upkeep's cascade can complete (Active investigator exists).
+        let mut events = Vec::new();
+
+        step_phase(&mut state, &mut events); // Enemy → Upkeep
+
+        // step_phase itself MUST NOT emit PhaseEnded(Enemy); only
+        // enemy_phase_end is allowed to (which doesn't run here — we
+        // started in Enemy and stepped out, simulating the "phase
+        // transition without driver-owned end emit" path).
+        let phase_ended_enemy_count = events
+            .iter()
+            .filter(|e| {
+                matches!(
+                    e,
+                    Event::PhaseEnded {
+                        phase: Phase::Enemy
+                    }
+                )
+            })
+            .count();
+        assert_eq!(
+            phase_ended_enemy_count, 0,
+            "step_phase must NOT emit PhaseEnded(Enemy); only enemy_phase_end may. events = {events:?}"
+        );
+    }
+
+    #[test]
+    fn enemy_phase_resumes_via_skip_input() {
+        // Construct the state mid-pause: a BeforeInvestigatorAttacked
+        // window is on the stack with empty pending_triggers (the
+        // "pure-Fast window" shape that open_fast_window pushes when
+        // Fast play is eligible), and the cursor points at inv1.
+        //
+        // Submitting PlayerAction::ResolveInput(InputResponse::Skip)
+        // routes through resolve_input's "open_windows non-empty +
+        // no reaction triggers" branch → close_reaction_window_at →
+        // run_window_continuation's BeforeInvestigatorAttacked arm →
+        // resolve_attacks_for_investigator → cursor advance to None →
+        // open AfterAllInvestigatorsAttacked → auto-skip continuation
+        // → enemy_phase_end → cascade Upkeep → Mythos.
+        //
+        // The synthetic OpenWindow push fakes the pause point because
+        // a real Fast-eligibility setup would require either a card-
+        // registry install (heavyweight integration test) or a Fast
+        // event card in hand with resources — neither tractable in
+        // the engine layer. The Skip path itself is the load-bearing
+        // resume mechanism this test exercises.
+        let inv_id = InvestigatorId(1);
+        let enemy_id = EnemyId(1);
+        let mut enemy = test_enemy(1, "Test Enemy");
+        enemy.engaged_with = Some(inv_id);
+        enemy.attack_damage = 1;
+        let mut state = TestGame::default()
+            .with_investigator(test_investigator(1))
+            .with_enemy(enemy)
+            .with_phase(Phase::Enemy)
+            .build();
+        state.turn_order = vec![inv_id];
+        state.active_investigator = None;
+        state.enemy_attack_pending = Some(inv_id);
+        state.open_windows.push(OpenWindow {
+            kind: WindowKind::BeforeInvestigatorAttacked,
+            pending_triggers: Vec::new(),
+            fast_actors: FastActorScope::Any,
+        });
+
+        let result = apply(
+            state,
+            Action::Player(PlayerAction::ResolveInput {
+                response: InputResponse::Skip,
+            }),
+        );
+
+        match result.outcome {
+            EngineOutcome::Done => {}
+            ref other => panic!(
+                "expected Done after Skip; got {other:?}; events = {:?}",
+                result.events
+            ),
+        }
+        assert_eq!(
+            result.state.phase,
+            Phase::Mythos,
+            "cascade lands in Mythos after Skip resumed the continuation"
+        );
+        assert!(
+            result.events.iter().any(|e| matches!(
+                e,
+                Event::DamageTaken { investigator, amount: 1 } if *investigator == inv_id
+            )),
+            "attack should have landed during the resumed continuation; events = {:?}",
+            result.events
+        );
+        assert!(
+            result.events.iter().any(|e| matches!(
+                e,
+                Event::EnemyExhausted { enemy } if *enemy == enemy_id
+            )),
+            "EnemyExhausted should fire during the resumed continuation; events = {:?}",
+            result.events
+        );
+        assert_eq!(
+            result.state.enemy_attack_pending, None,
+            "cursor must clear after the continuation advances past the last \
+             Active investigator and the AfterAll window auto-skips"
+        );
+    }
+
+    // TODO(#71 follow-up): pause-on-Fast-eligibility test — needs a
+    // tractable Fast-eligibility fixture at the engine layer (Fast
+    // event card in hand + resources + card-registry install, which
+    // would push this into the cards crate's integration tests). The
+    // Skip-resume test above proves the resume path is correct; the
+    // pause shape is exercised indirectly via the existing
+    // any_fast_play_eligible-driven open_fast_window tests at
+    // dispatch.rs's open_fast_window_tests block.
+}
